@@ -15,13 +15,15 @@
     "@choc-ui/logo": "^1.0.0",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
+    "@reduxjs/toolkit": "^2.2.6",
     "framer-motion": "^5.6.0",
     "leaflet": "^1.7.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^4.3.1",
     "react-leaflet": "^3.2.5",
     "react-query": "^3.34.12",
+    "react-redux": "^9.1.2",
     "react-router-dom": "^6.2.1",
     "webpack-dev-server": "^4.7.4"
   },
