@@ -1,8 +1,10 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import App from './App';
 import { QueryClientProvider, QueryClient } from 'react-query';
 const queryClient = new QueryClient();
 
-ReactDOM.render(<App />, document.querySelector('#root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App tab="home" />);
