@@ -1,5 +1,6 @@
 import { doUntil } from 'async';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect, useState, useContext } from 'react';
+import { Provider } from 'react-redux';
 import {
 	Text,
 	HStack,
@@ -12,22 +13,40 @@ import NavbarComponent from './components/Navbar';
 import MTARoutesComponent from './components/MTARoutes';
 import LeafletMapComponent from './components/LeafletMap';
 import { QueryClientProvider, QueryClient } from 'react-query';
+import store from './store';
 import theme from './theme';
+import { setRoutes } from './actions/routes';
 
 const queryClient = new QueryClient();
 
 export default function App({ Component }) {
+	useEffect(() => {
+		// TODO: #5 - fix race conditions
+		// TODO: This should likely be set in thunk, saga, or use middleware api
+
+		fetch('/api/routes')
+			.then((res) => res.json())
+			.then((routes) => store.dispatch(setRoutes(routes)))
+			.catch((e) =>
+				store.dispatch({ type: 'SET_ROUTES', payload: routes })
+			);
+	}, []);
+
+	const state = store.getState();
+
 	return (
 		<QueryClientProvider client={queryClient}>
 			<ChakraProvider theme={theme}>
-				<NavbarComponent />
-				<LeafletMapComponent
-					h={'250px'}
-					w={'100%'}
-					bg="black"
-					position={[40.7812, -73.9665]}
-				/>
-				<MTARoutesComponent />
+				<Provider store={store}>
+					<NavbarComponent />
+					<LeafletMapComponent
+						h={'250px'}
+						w={'100%'}
+						bg="black"
+						position={[40.7812, -73.9665]}
+					/>
+					<MTARoutesComponent />
+				</Provider>
 			</ChakraProvider>
 		</QueryClientProvider>
 	);
